+++ mercado-site-retriever/src/main/java/org/marceloleite/mercado/siteretriever/util/converter/ListToMapJsonTradeConverter.java
@@ -7,10 +7,10 @@ import java.util.stream.Collectors;
 import org.marceloleite.mercado.commons.util.converter.Converter;
 import org.marceloleite.mercado.siteretriever.model.JsonTrade;
 
-public class ListToMapJsonTradeConverter implements Converter<List<JsonTrade>, Map<Integer, JsonTrade>> {
+public class ListToMapJsonTradeConverter implements Converter<List<JsonTrade>, Map<Long, JsonTrade>> {
 
 	@Override
-	public Map<Integer, JsonTrade> convert(List<JsonTrade> jsonTrades) {
+	public Map<Long, JsonTrade> convert(List<JsonTrade> jsonTrades) {
 		return jsonTrades.stream()
 			.map(jsonTrade -> jsonTrade)
 			.collect(Collectors.toConcurrentMap(JsonTrade::getTid, jsonTrade -> jsonTrade, (oldJsonTrade, newJsonTrade) -> newJsonTrade));
